// frontend/src/lib/api/types.gen.ts
export type Network = {
    machineId?: string;
    name: string;
    network?: string;
    driver: string;
    status: string;
};

export type Volume = {
    id?: string;
    name: string;
    driver: string;
    status: string;
    source?: string;
    destination?: string;
};